+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 Yii2 multiple input change log
 ==============================
 
-1.2.20 in development
----------------------
+1.3.1 in development
+--------------------
+
+1.3.0
+-----
 
 - #79 Added support for embedded MultipleInput widget (unclead, execut)
 - Enh: Added ability to render `add` button in the footer (unclead)